This patch fixes the failing case reported, but is really a work around of a considerably deeper bug.
When GetEndOffsetFromObj() needs to do index arithmetic calculations with
bignum arguments, rather than code all of that directly, it tries to call
on the routine we already have for that, Tcl_ExprObj().

Problem is that apparently a call to Tcl_ExprObj() in certain scenarios leads
to nested (?) calls of TclNRExecuteByteCode() and somehow this mangles the
expectations of the TclStack(Alloc|Free) routines in their abuse of the
execution stack space.

It's very complicated and I haven't figured it out yet.

We can accept this patch as a fix for the reported bug.  But if so, we should
open a new bug that Tcl_ExprObj() has undocumented conditions on when it
can be safely called.  That's bad for a public routine.